pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tower-tools"
version = "0.0.1"
description = "Tower tools"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "gevent",
    "x2go @ https://code.x2go.org/releases/source/python-x2go/python-x2go-0.6.1.3.tar.gz",
    "python-xlib==0.33",
    "requests==2.27.1",
    "sh==1.14.3",
    "backports.pbkdf2==0.1",
    "passlib==1.7.4",
    "sshconf==0.2.5",
    #"tower-tools[tower]",
]

[tool.hatch.build.targets.wheel]
packages = ["tower"]

[tool.hatch.build]
include = [
  "tower/raspberrypios/*.patch",
]

[project.scripts]
tower = "tower.clis.tower:main"
build-tower-image = "tower.clis.buildimage:main"

[tool.hatch.metadata]
allow-direct-references=true

#[tool.hatch.envs.default]
#python = "/opt/homebrew/opt/python@3.10/bin/python3.10"